Chore(Gdacs): handle impact data in extraction phase
