+++ apps/etl/extraction/sources/gdacs/base.py
@@ -1,0 +1,9 @@
+from enum import Enum
+
+
+class GdacsExtractionMetadataType(str, Enum):
+    QUERY = "QUERY"
+    DETAIL = "DETAIL"
+    GEOMETRY = "GEOMETRY"
+    EPISODE = "EPISODE"
+    IMPACT = "IMPACT"
